+++ Dockerfile
@@ -3,7 +3,7 @@
 #
 
 # Uses Ubuntu 16.04 LTS
-FROM flywheel/hcp-base:0.1.0_4.0.1
+FROM flywheel/hcp-base:0.2.0_4.0.1
 
 LABEL maintainer="Flywheel <support@flywheel.io>"
 
@@ -11,15 +11,15 @@ LABEL maintainer="Flywheel <support@flywheel.io>"
 # FSL 6.0.1 is a part of the base image.  Update the environment variables
 
 # Configure FSL environment
-ENV FSLDIR=/usr/share/fsl/6.0
-ENV FSL_DIR="${FSLDIR}"
-ENV FSLOUTPUTTYPE=NIFTI_GZ
-ENV PATH=/usr/share/fsl/6.0/bin:$PATH
-ENV FSLMULTIFILEQUIT=TRUE
-ENV POSSUMDIR=/usr/share/fsl/6.0
-ENV LD_LIBRARY_PATH=/usr/share/fsl/6.0/lib:$LD_LIBRARY_PATH
-ENV FSLTCLSH=/usr/bin/tclsh
-ENV FSLWISH=/usr/bin/wish
+ENV FSLDIR=/usr/share/fsl/6.0 \ 
+    FSL_DIR="${FSLDIR}" \ 
+    FSLOUTPUTTYPE=NIFTI_GZ \ 
+    PATH=/usr/share/fsl/6.0/bin:$PATH \ 
+    FSLMULTIFILEQUIT=TRUE \ 
+    POSSUMDIR=/usr/share/fsl/6.0 \ 
+    LD_LIBRARY_PATH=/usr/share/fsl/6.0/lib:$LD_LIBRARY_PATH \ 
+    FSLTCLSH=/usr/bin/tclsh \ 
+    FSLWISH=/usr/bin/wish
 
 #############################################
 # Download and install Connectome Workbench 1.3.2 
@@ -33,34 +33,32 @@ ENV CARET7DIR=/opt/workbench/bin_linux64
 # Using v4.0.1
 
 # Set up specific environment variables for the HCP Pipeline
-ENV FSL_DIR="${FSLDIR}"
-ENV HCPPIPEDIR=/opt/HCP-Pipelines
-ENV MSMBINDIR=${HCPPIPEDIR}/MSMBinaries
-ENV MSMCONFIGDIR=${HCPPIPEDIR}/MSMConfig
+ENV FSL_DIR="${FSLDIR}" \ 
+    HCPPIPEDIR=/opt/HCP-Pipelines \ 
+    MSMBINDIR=${HCPPIPEDIR}/MSMBinaries \ 
+    MSMCONFIGDIR=${HCPPIPEDIR}/MSMConfig
 #ENV MATLAB_COMPILER_RUNTIME=/media/myelin/brainmappers/HardDrives/1TB/MATLAB_Runtime/v901
 #ENV FSL_FIXDIR=/media/myelin/aahana/fix1.06
 
-#For HCP Pipeline v3.x
-ENV MSMBin=${HCPPIPEDIR}/MSMBinaries
-
-ENV HCPPIPEDIR_Templates=${HCPPIPEDIR}/global/templates
-ENV HCPPIPEDIR_Bin=${HCPPIPEDIR}/global/binaries
-ENV HCPPIPEDIR_Config=${HCPPIPEDIR}/global/config
-
-ENV HCPPIPEDIR_PreFS=${HCPPIPEDIR}/PreFreeSurfer/scripts
-ENV HCPPIPEDIR_FS=${HCPPIPEDIR}/FreeSurfer/scripts
-ENV HCPPIPEDIR_PostFS=${HCPPIPEDIR}/PostFreeSurfer/scripts
-ENV HCPPIPEDIR_fMRISurf=${HCPPIPEDIR}/fMRISurface/scripts
-ENV HCPPIPEDIR_fMRIVol=${HCPPIPEDIR}/fMRIVolume/scripts
-ENV HCPPIPEDIR_tfMRI=${HCPPIPEDIR}/tfMRI/scripts
-ENV HCPPIPEDIR_dMRI=${HCPPIPEDIR}/DiffusionPreprocessing/scripts
-ENV HCPPIPEDIR_dMRITract=${HCPPIPEDIR}/DiffusionTractography/scripts
-ENV HCPPIPEDIR_Global=${HCPPIPEDIR}/global/scripts
-ENV HCPPIPEDIR_tfMRIAnalysis=${HCPPIPEDIR}/TaskfMRIAnalysis/scripts
+#For HCP Pipeline v4.0.1
+ENV MSMBin=${HCPPIPEDIR}/MSMBinaries \
+    HCPPIPEDIR_Templates=${HCPPIPEDIR}/global/templates \ 
+    HCPPIPEDIR_Bin=${HCPPIPEDIR}/global/binaries \ 
+    HCPPIPEDIR_Config=${HCPPIPEDIR}/global/config \ 
+    HCPPIPEDIR_PreFS=${HCPPIPEDIR}/PreFreeSurfer/scripts \ 
+    HCPPIPEDIR_FS=${HCPPIPEDIR}/FreeSurfer/scripts \ 
+    HCPPIPEDIR_PostFS=${HCPPIPEDIR}/PostFreeSurfer/scripts \ 
+    HCPPIPEDIR_fMRISurf=${HCPPIPEDIR}/fMRISurface/scripts \ 
+    HCPPIPEDIR_fMRIVol=${HCPPIPEDIR}/fMRIVolume/scripts \ 
+    HCPPIPEDIR_tfMRI=${HCPPIPEDIR}/tfMRI/scripts \ 
+    HCPPIPEDIR_dMRI=${HCPPIPEDIR}/DiffusionPreprocessing/scripts \ 
+    HCPPIPEDIR_dMRITract=${HCPPIPEDIR}/DiffusionTractography/scripts \ 
+    HCPPIPEDIR_Global=${HCPPIPEDIR}/global/scripts \ 
+    HCPPIPEDIR_tfMRIAnalysis=${HCPPIPEDIR}/TaskfMRIAnalysis/scripts
 
 #try to reduce strangeness from locale and other environment settings
-ENV LC_ALL=C
-ENV LANGUAGE=C
+ENV LC_ALL=C \ 
+    LANGUAGE=C
 #POSIXLY_CORRECT currently gets set by many versions of fsl_sub, unfortunately, but at least don't pass it in if the user has it set in their usual environment
 RUN unset POSIXLY_CORRECT
 
@@ -70,23 +68,22 @@ RUN unset POSIXLY_CORRECT
 # 5.3.0 ftp://surfer.nmr.mgh.harvard.edu/pub/dist/freesurfer/5.3.0-HCP/freesurfer-Linux-centos4_x86_64-stable-pub-v5.3.0-HCP.tar.gz
 
 # Set up the FreeSurfer environment
-ENV OS Linux
-ENV FS_OVERRIDE 0
-ENV FIX_VERTEX_AREA=
-ENV SUBJECTS_DIR /opt/freesurfer/subjects
-ENV FSF_OUTPUT_FORMAT nii.gz
-ENV MNI_DIR /opt/freesurfer/mni
-ENV LOCAL_DIR /opt/freesurfer/local
-ENV FREESURFER_HOME /opt/freesurfer
-ENV FSFAST_HOME /opt/freesurfer/fsfast
-ENV MINC_BIN_DIR /opt/freesurfer/mni/bin
-ENV MINC_LIB_DIR /opt/freesurfer/mni/lib
-ENV MNI_DATAPATH /opt/freesurfer/mni/data
-ENV FMRI_ANALYSIS_DIR /opt/freesurfer/fsfast
-ENV PERL5LIB /opt/freesurfer/mni/lib/perl5/5.8.5
-ENV MNI_PERL5LIB /opt/freesurfer/mni/lib/perl5/5.8.5
-ENV PATH /opt/freesurfer/bin:/opt/freesurfer/fsfast/bin:/opt/freesurfer/tktools:/opt/freesurfer/mni/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:$PATH
-
+ENV OS=Linux \ 
+    FS_OVERRIDE=0 \ 
+    FIX_VERTEX_AREA= \ 
+    SUBJECTS_DIR=/opt/freesurfer/subjects \ 
+    FSF_OUTPUT_FORMAT=nii.gz \ 
+    MNI_DIR=/opt/freesurfer/mni \ 
+    LOCAL_DIR=/opt/freesurfer/local \ 
+    FREESURFER_HOME=/opt/freesurfer \ 
+    FSFAST_HOME=/opt/freesurfer/fsfast \ 
+    MINC_BIN_DIR=/opt/freesurfer/mni/bin \ 
+    MINC_LIB_DIR=/opt/freesurfer/mni/lib \ 
+    MNI_DATAPATH=/opt/freesurfer/mni/data \ 
+    FMRI_ANALYSIS_DIR=/opt/freesurfer/fsfast \ 
+    PERL5LIB=/opt/freesurfer/mni/lib/perl5/5.8.5 \ 
+    MNI_PERL5LIB=/opt/freesurfer/mni/lib/perl5/5.8.5 \ 
+    PATH=/opt/freesurfer/bin:/opt/freesurfer/fsfast/bin:/opt/freesurfer/tktools:/opt/freesurfer/mni/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:$PATH
 
 #############################################
 # Gradient unwarp script is installed in base image. 
@@ -97,9 +94,9 @@ ENV MSMBINDIR=${HCPPIPEDIR}/MSMBinaries
 
 #############################################
 # Copy gear-specific utils and dependencies
-COPY struct_utils ${FLYWHEEL}/utils
-COPY struct_scripts /tmp/scripts
-COPY struct_scenes /tmp/scenes
+COPY utils ${FLYWHEEL}/utils
+COPY scripts /tmp/scripts
+COPY scenes /tmp/scenes
 
 # Copy executable/manifest to Gear
 COPY run.py ${FLYWHEEL}/run.py
