+++ run.py
@@ -6,53 +6,44 @@ import copy
 import shutil
 
 import flywheel
-from utils.args.Common import set_subject
-from utils.custom_logger import get_custom_logger, log_config
+from utils.custom_logger import get_custom_logger
 from utils.args import PreFreeSurfer, FreeSurfer, PostFreeSurfer
 from utils.args import hcpstruct_qc_scenes, hcpstruct_qc_mosaic
 from utils.args import PostProcessing
 from utils import results, gear_preliminaries
+from utils.struct_utils import get_freesurfer_version
 
 if __name__ == '__main__':
-    # Get the Gear Context
+    # Preamble: take care of all gear-typical activities.
     context = flywheel.GearContext()
-    # Activate custom logger
-    context.log = get_custom_logger('[flywheel/hcp-struct]')
+    context.gear_dict = {}
+    # Initialize all hcp-gear variables.
+    gear_preliminaries.initialize_gear(context)
+    context.log_config()
 
     # Validate gear configuration against gear manifest
     try:
         gear_preliminaries.validate_config_against_manifest(context)
     except Exception as e:
-        context.log.fatal(e,)
+        context.log.exception(e)
         context.log.fatal(
             'Please make the prescribed corrections and try again.'
         )
         os.sys.exit(1)
-    # Set up Custom Dicionary to host user variables
-    context.gear_dict={}
-    context.gear_dict['SCRIPT_DIR']    = '/tmp/scripts'
-    context.gear_dict['SCENE_DIR']     = '/tmp/scenes'
-    # Can I automate this? Do I want to?
-    context.gear_dict['FreeSurfer_Version'] = '5.3.0'
-    # Instantiate Environment Variables
-    # This will always be '/tmp/gear_environ.json' with these 
-    # environments defined in the Dockerfile and exported from there.
-    with open('/tmp/gear_environ.json', 'r') as f:
-        environ = json.load(f)
 
-    context.gear_dict['environ'] = environ
-    # Create a 'dry run' flag for debugging
-    context.gear_dict['dry-run'] = context.config['Dry-Run']
+    # Can I automate this? Do I want to?
+    context.gear_dict['FreeSurfer_Version'] = get_freesurfer_version(context)
+
      
     ###########################################################################
     # Pipelines common commands
-    QUEUE = ""
+    # QUEUE works differently in FSL 6.0.1..we are not using it.
+    QUEUE = "-q "
     LogFileDirFull = op.join(context.work_dir,'logs')
     os.makedirs(LogFileDirFull, exist_ok=True)
     FSLSUBOPTIONS = "-l "+ LogFileDirFull
-
-    command_common=[op.join(environ['FSLDIR'],'bin','fsl_sub'),
-                   QUEUE, FSLSUBOPTIONS]
+    environ = context.gear_dict['environ']
+    command_common=[op.join(environ['FSLDIR'],'bin','fsl_sub'), FSLSUBOPTIONS]
     
     context.gear_dict['command_common'] = command_common
 
@@ -64,22 +55,20 @@ if __name__ == '__main__':
     ###########################################################################
     # Ensure the subject_id is set in a valid manner (api or config)
     try:
-        set_subject(context)
+        gear_preliminaries.set_subject(context)
     except Exception as e:
-        context.log.fatal(e,)
+        context.log.exception(e)
         context.log.fatal(
             'The Subject ID is not valid. Examine and try again.',
         )
         os.sys.exit(1)
 
-    # Report on Inputs and configuration parameters to the log
-    log_config(context)
     # Build and Validate Parameters for the PreFreeSurferPipeline.sh 
     try:
         PreFreeSurfer.build(context)
         PreFreeSurfer.validate(context)
     except Exception as e:
-        context.log.fatal(e,)
+        context.log.exception(e)
         context.log.fatal(
             'Validating Parameters for the PreFreeSurferPipeline Failed.',
         )
@@ -93,7 +82,7 @@ if __name__ == '__main__':
         # No user-submitted parameters to validate at this level
         # FreeSurfer.validate(context)
     except Exception as e:
-        context.log.fatal(e)
+        context.log.exception(e)
         context.log.fatal(
             'Validating Parameters for the FreeSurferPipeline Failed.'
         )
@@ -105,7 +94,7 @@ if __name__ == '__main__':
         PostFreeSurfer.build(context)
         PostFreeSurfer.validate(context)
     except Exception as e:
-        context.log.fatal(e)
+        context.log.exception(e)
         context.log.fatal(
             'Validating Parameters for the PostFreeSurferPipeline Failed!'
         )
@@ -116,7 +105,7 @@ if __name__ == '__main__':
     try:
         PreFreeSurfer.execute(context)
     except Exception as e:
-        context.log.fatal(e,)
+        context.log.exception(e)
         context.log.fatal(
             'The PreFreeSurferPipeline Failed.',
         )
@@ -130,7 +119,7 @@ if __name__ == '__main__':
         FreeSurfer.validate(context)
         FreeSurfer.execute(context)
     except Exception as e:
-        context.log.fatal(e)
+        context.log.exception(e)
         context.log.fatal('The FreeSurferPipeline Failed.')
         if context.config['save-on-error']:
             results.cleanup(context)
@@ -141,7 +130,7 @@ if __name__ == '__main__':
     try:
         PostFreeSurfer.execute(context)
     except Exception as e:
-        context.log.fatal(e)
+        context.log.exception(e)
         context.log.fatal('The PostFreeSurferPipeline Failed!')
         if context.config['save-on-error']:
             results.cleanup(context)
@@ -155,7 +144,7 @@ if __name__ == '__main__':
         PostProcessing.build(context)
         PostProcessing.execute(context)
     except Exception as e:
-        context.log.fatal(e)
+        context.log.exception(e)
         context.log.fatal('The Post Processing Failed!')
         if context.config['save-on-error']:
             results.cleanup(context)
@@ -170,7 +159,7 @@ if __name__ == '__main__':
         hcpstruct_qc_mosaic.build(context)
         hcpstruct_qc_mosaic.execute(context)
     except Exception as e:
-        context.log.fatal(e,)
+        context.log.exception(e)
         context.log.fatal('HCP Structural QC Images has failed!')
         if context.config['save-on-error']:
             results.cleanup(context)
