+++ scenes/TEMPLATE.hcpstruct_QC.very_inflated.164k_fs_LR.scene
