+++ scripts/hcpstruct_qc_mosaic.sh
