+++ scripts/hcpstruct_qc_scenes.sh
