+++ utils/struct_utils.py
